+++ tests/smart_contracts/test_forms.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+from typing import Type
 
 import pytest
 from django.contrib.contenttypes.models import ContentType
@@ -7,18 +8,19 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django_webtest import DjangoTestApp
 
 from tests.conftest import organization
+from tests.smart_contracts.conftest import agreement_pdf, ODRL_JSON
 from vitrina.datasets.factories import DatasetFactory, ContactFactory
 from vitrina.datasets.models import Dataset
 from vitrina.orgs.factories import OrganizationFactory
 from vitrina.orgs.models import Organization
 from vitrina.projects.factories import ProjectFactory
-from vitrina.smart_contracts.factories import AgreementPDFFileFactory, AgreementFactory
+from vitrina.smart_contracts import AgreementStatuses
+from vitrina.smart_contracts.factories import AgreementPDFFileFactory, AgreementFactory, AgreementJSONFileFactory
 from vitrina.smart_contracts.forms import (
     SmartContractForm,
-    AgreementUploadForm,
     AgreementSubmitForm,
     AgreementApproveForm,
-    AgreementFormForm,
+    AgreementFormForm, AgreementInitiateForm, AgreementSignForm,
 )
 from vitrina.smart_contracts.models import SmartContractTemplate
 from vitrina.structure.factories import MetadataFactory
@@ -297,7 +299,6 @@ class TestAgreementApproveForm:
         assert form.errors == {
             "template": ["Šis laukas yra privalomas."],
             "assigner_representative": ["Šis laukas yra privalomas."],
-            "other_assigner_legislations": ["Šis laukas yra privalomas."],
         }
 
 
@@ -316,20 +317,110 @@ class TestAgreementFormForm:
         assert form.is_valid(), form.errors
 
 
-class TestAgreementUploadForm:
-    def test_not_valid_when_uploading_file_other_than_adoc(self) -> None:
-        uploaded_file = SimpleUploadedFile("bad_file.md", b"md file content")
-        agreement_pdf = AgreementPDFFileFactory()
-        form = AgreementUploadForm(files={"file": uploaded_file}, agreement_pdf=agreement_pdf, agreement=agreement_pdf.agreement)
+class TestAgreementInitiateAndSignForms:
+    """Tests for both AgreementInitiateForm and AgreementSignForm have the same logic in their respective forms."""
+    @pytest.mark.parametrize(
+        "form_class,agreement_status,file_name",
+        [
+            (AgreementInitiateForm, AgreementStatuses.FORMED, "agreement_one_signer.adoc"),
+            (AgreementSignForm, AgreementStatuses.INITIATED, "agreement_two_signers.adoc")
+        ],
+    )
+    def test_success(
+        self,
+        form_class: Type[AgreementInitiateForm | AgreementSignForm],
+        agreement_status: AgreementStatuses,
+        file_name: str,
+        agreement_pdf: str
+    ):
+        # Arrange
+        base_path = Path(__file__).parent / "files" / "test_contracts"
+        test_file_path = base_path / file_name
+        odrl_json_file_path = Path(__file__).parent / "files" / ODRL_JSON
 
-        assert form.is_valid() is False
+        with open(test_file_path, "rb") as file:
+            uploaded_file = SimpleUploadedFile("agreement_one_signer.adoc", file.read())
+        agreement = AgreementFactory(status=agreement_status)
+        agreement_pdf_file = AgreementPDFFileFactory(pdf_path=agreement_pdf, agreement=agreement)
+
+        AgreementJSONFileFactory(agreement=agreement, json_path=odrl_json_file_path)
+
+        # Act
+        form = form_class(
+            files={"file": uploaded_file},
+            agreement_pdf=agreement_pdf_file,
+            agreement=agreement,
+        )
+
+        # Assert
+        assert form.is_valid()
+
+    def test_uploaded_document_is_not_adoc(self):
+        # Arrange
+        uploaded_file = SimpleUploadedFile("not_adoc.md", b"content")
+        agreement_pdf_file = AgreementPDFFileFactory()
+
+        # Act
+        form = AgreementInitiateForm(files={"file": uploaded_file}, agreement_pdf=agreement_pdf_file, agreement=agreement_pdf_file.agreement)
+
+        # Assert
+        assert not form.is_valid()
         assert form.errors == {"file": ["Dokumentas turi būti adoc formato."]}
 
-    def test_not_valid_when_uploading_unsigned_adoc(self, agreement_pdf: Path, agreement_not_signed: Path) -> None:
-        agreement_pdf = AgreementPDFFileFactory(pdf_path = agreement_pdf)
-        with open(agreement_not_signed, "rb") as f:
-            uploaded_file = SimpleUploadedFile("sutartis_not_signed.adoc", f.read())
+    def test_bad_zip_file(self):
+        # Arrange
+        uploaded_file = SimpleUploadedFile("bad.adoc", b"content")
+        agreement_pdf_file = AgreementPDFFileFactory()
 
-        form = AgreementUploadForm(files={"file": uploaded_file}, agreement_pdf=agreement_pdf, agreement=agreement_pdf.agreement)
-        assert form.is_valid() is False
-        assert form.errors == {"file": ["Įkelta sutartis nepasirašyta."]}
+        # Act
+        form = AgreementInitiateForm(
+            files={"file": uploaded_file},
+            agreement_pdf=agreement_pdf_file,
+            agreement=agreement_pdf_file.agreement
+        )
+
+        # Assert
+        assert not form.is_valid()
+        assert form.errors == {"file": ["Prisegtas failas nėra ZIP archyvas."]}
+
+    @pytest.mark.parametrize("form_class", [AgreementInitiateForm, AgreementSignForm])
+    @pytest.mark.parametrize(
+        "filename,expected_error",
+        [
+            ("agreement_bad_certificate.adoc", "ADOC klaida: Netinkamas parašo sertifikatas."),
+            ("agreement_no_manifest.adoc", "ADOC klaida: Neteisingas ADOC formatas."),
+            ("agreement_two_files.adoc", "ADOC klaida: Rastas daugiau nei vienas pasirašytas dokumentas."),
+            ("agreement_no_pdf.adoc", "ADOC klaida: Nerastas PDF dokumentas."),
+            ("agreement_modified.adoc", "ADOC klaida: PDF dokumentas nesutampa su sutartyje esančiu PDF dokumentu."),
+            ("agreement_non_zip.adoc", "Prisegtas failas nėra ZIP archyvas."),
+            ("agreement_not_signed.adoc", "Įkelta sutartis nepasirašyta."),
+            ("agreement.pdf", "Dokumentas turi būti adoc formato."),
+        ]
+    )
+    def test_initiate_and_sign_form_errors(
+        self,
+        form_class: Type[AgreementInitiateForm | AgreementSignForm],
+        filename: str,
+        expected_error: str,
+        agreement_pdf: str,
+    ):
+        """Checks both forms and all validations, that are common throughout these forms."""
+        # Arrange
+        base_path = Path(__file__).parent / "files" / "test_contracts"
+        test_file_path = base_path / filename
+
+        with open(test_file_path, "rb") as f:
+            uploaded_file = SimpleUploadedFile(filename, f.read())
+        agreement_pdf_file = AgreementPDFFileFactory(pdf_path=agreement_pdf)
+
+        # Act
+        form = form_class(
+            files={"file": uploaded_file},
+            agreement_pdf=agreement_pdf_file,
+            agreement=agreement_pdf_file.agreement,
+        )
+
+        # Assert
+        assert not form.is_valid()
+        assert "file" in form.errors
+        assert form.errors["file"][0] == expected_error