+++ vitrina/users/models.py
@@ -49,12 +49,18 @@ class User(AbstractUser):
     )
     failed_login_attempts = models.IntegerField(default=0)
     password_last_updated = models.DateTimeField(default=now, null=True)
-    is_viisp_login = models.BooleanField(default=False)
+    is_viisp_login = models.BooleanField(
+        verbose_name=_("Vartotojas prisijungęs per VIISP"),
+        default=False,
+        help_text=_("Indikuoja, ar vartotojas yra prisijungęs per VIISP."),
+    )
     viisp_company_code = models.CharField(
+        verbose_name=_("Vartotojo atstovaujama organizacija (VIISP)"),
         max_length=255,
         blank=True,
         null=True,
         editable=False,
+        help_text=_("Nurodomas įmonės kodas organizacijos, kurios vardu vartotojas yra prisijungęs per VIISP."),
     )
     receive_request_email = models.BooleanField(default=False)